refactor: Consolidate duplicate helper functions (CLD-001) and remove debug statements
## Architectural Improvements

### Helper Function Consolidation (CLD-001)
- Moved safe_float() and sanitize_str() to _utils.py
- Removed 3 duplicate implementations across modules
- Reduced code duplication by ~32 lines
- Single source of truth for critical parsing utilities

### Debug Statement Cleanup
- Removed verbose text content logging from _query.py
- Retained entity ranking and count statistics
- Improved security by not exposing text content in logs

## Files Changed
- nano_graphrag/_utils.py: Added centralized helper functions
- nano_graphrag/_extraction.py: Removed duplicate helpers, updated imports
- nano_graphrag/entity_extraction/llm.py: Removed duplicate helpers, updated imports
- nano_graphrag/_query.py: Removed debug statements showing text content

## Benefits
- Improved maintainability with single implementation
- Consistent behavior across all modules
- Cleaner logs without sensitive content exposure
- Easier testing and bug fixes

All tests passing (378 tests) - no functional changes.
