+++ documentation/reviews/NGRAF-020-claude-round7.md
@@ -1,0 +1,19 @@
+# Abstract
+The submission completes the NGRAF-020 / NGRAF-020-2 migrations by standardising entity extraction on NDJSON, repairing sparse-vector preservation, and hardening the Qdrant integration. Architecturally, the new workflow is coherent and well tested, but duplicated parsing utilities across modules pose a maintainability risk that should be consolidated before merge. No blocking defects were found.
+
+## High Priority Issues
+_None._
+
+## Medium Priority Issues
+- **CLD-001**: `nano_graphrag/_extraction.py:266-338`, `nano_graphrag/entity_extraction/llm.py:132-199`, `nano_graphrag/_extraction.py:475-520` | Medium | Each module hand-rolls nearly identical `sanitize_str` / `safe_float` helpers and NDJSON parsing loops. Any schema change (e.g. new fields, stricter sanitisation) must now be synchronised in three places, inviting drift and future regressions. | Extract the shared helpers into a dedicated utility (e.g. `nano_graphrag._utils.parse_ndjson_record` plus `sanitize_str` / `safe_float`) and call it from both legacy and new extractor paths.
+
+## Low Priority Notes
+- **CLD-OBS-001**: `nano_graphrag/_storage/vdb_qdrant.py:162-214` | The sparse-name channel is now produced alongside the primary sparse embedding. Consider documenting the expectation that both arrays stay in lock-step and adding assertions to catch SPLADE responses of mismatched length.
+
+## Positive Observations
+- **CLD-GOOD-001**: `nano_graphrag/prompt.py:200-320` | Prompt examples and instructions now accurately describe the NDJSON contract, dramatically reducing ambiguity for model outputs.
+- **CLD-GOOD-002**: `tests/test_ndjson_extraction.py:1-220` | Comprehensive regression suite covering null fields, newline handling, and LLM extractor integration provides confidence in the new parser.
+- **CLD-GOOD-003**: `nano_graphrag/graphrag.py:520-541` | Community fallback now mirrors the initial upsert payload, preserving SPLADE signal and preventing the sparse-vector regression reported by the product owner.
+
+## Recommendations
+Resolve CLD-001 by centralising NDJSON parsing helpers, then proceed with merge. Optionally add length assertions in the Qdrant layer as noted in CLD-OBS-001 for defensive robustness.
